Ament the centralised status metadata utility
